feat!: migrate to flat config
BREAKING CHANGES:

- Requires ESLint 9.33.0 or later.
- Dropped `.eslintrc*` support. Only available as flat config now.
- Updated to @antfu/eslint-config from 0.41.0 to 5.2.1, so a lot of
rules have changed. I tried to keep most of the rules the same, but
there are still some minor differences. Most are space-related stylistic
ones and are auto-fixable.

A few things that need to be discussed:

- As `@antfu/eslint-config` [does not follow semver strictly](https://github.com/antfu/eslint-config/tree/v5.2.1?tab=readme-ov-file#versioning-policy),
do we need to pin the dependency version?
- Antfu's config renamed the `@typescript-eslint/` prefix to `ts/`,
I chose to follow that, because it's more concise. This affects how we
override rules in our codebase, but that's a rare case. And as I've
checked, many are references to `@typescript-eslint/ban-types`, which is
a [deprecated rule](https://typescript-eslint.io/rules/ban-types/) that
needs migration anyway.
- Other rules requiring discussion are noted in code comments.

Closes #34